+++ CHANGES.rst
@@ -66,6 +66,10 @@ astropy.modeling
 astropy.nddata
 ^^^^^^^^^^^^^^
 
+- Added support in the ``bitmask`` module for using mnemonic bit flag names
+  when specifying the bit flags to be used or ignored when converting a bit
+  field to a boolean. [#10095]
+
 astropy.samp
 ^^^^^^^^^^^^